Check if user with omniauth email exists
When a user try to access redmine from saml we try to create it if ti
doesn't exists a user with specified login name. That was not enough
because, there may be a user with other login name in redmine but with
the same email that comes from saml. So we try to find that user and
return it as logged in user